refactor(#7): integrate global exceptions, added a new bad request exception ,configured test cases into seperate classes and addressed PR feedback
